Refactor: PostCard 컴포넌트 수정 #2
