+++ components/pages/main/Feed/PostsList/FeedPostCard.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, SetStateAction } from 'react';
 import { Comment } from '@/types/types';
-import PostCard from '.';
-import FlexBox from '../FlexBox';
+import { PostCard } from '../../../../ui/PostCard';
+import FlexBox from '../../../../ui/FlexBox';
 
 interface FeedPostCardProps {
   userId: number;
